Refactor date parse tests to remove the duplication among the 3 files. We are going to be adding a bunch more tests to fix lots of bugs with date parsing. It is difficult to remember to put the same test in multiple differnt places each time you add a scenario. Also it gets hard to track the duplication across the various classes. Lots of the date tests are identical to DateTime tests. I think a future refactor might be able to figure out how to reuse the Date tests and the DateTime tests can just focus on things that include time
